style: update setting ui
